+++ CHANGELOG.md
@@ -21,7 +21,11 @@ and start a new "In Progress" section above it.
 
 ## In progress
 
+
+## 0.127.0
+
 - Add `simple_job_progress_estimation` config for simple job progress estimation ([Open-EO/openeo-geopyspark-driver#772](https://github.com/Open-EO/openeo-geopyspark-driver/issues/772))
+- `OpenEoBackendConfig`: be more forgiving about unknown config keys to better support use cases that involve backward/forward incompatible configurations ([#322](https://github.com/Open-EO/openeo-python-driver/issues/322))
 
 
 ## 0.126.0